backend/requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
google-generativeai
python-dotenv